test: Fix failing baudrate conversion
The definition of `speed_t` varies between versions of the libc. In
particular, the latest glibc versions alias it to `unsigned int` and the
`B*` macros behave like enum variants. That is not the case for the
target of Beluga (glibc v2.34), where `speed_t` is `long int` and the
macros behave like constants.